Move repo module to src/repo/
