+++ owners/test/parser.test.js
@@ -19,7 +19,7 @@ const path = require('path');
 const fs = require('fs');
 const JSON5 = require('json5');
 const {Team} = require('../src/github');
-const {LocalRepository} = require('../src/repo');
+const {LocalRepository} = require('../src/repo/repo');
 const {OwnersParser, OwnersParserError} = require('../src/parser');
 const {
   UserOwner,
